requirements.txt
# requirements.txt
# Surgical Risk Prediction Project - Required Libraries with Tested Versions
# Tested on Python 3.13.5, macOS with Apple Silicon
# Last updated: November 13, 2025

# ============================================================================
# Core Data Processing
# ============================================================================
numpy==2.1.3
pandas==2.2.3
scipy==1.15.3

# ============================================================================
# Machine Learning
# ============================================================================
scikit-learn==1.6.1
joblib==1.4.2

# ============================================================================
# Deep Learning - PyTorch
# ============================================================================
torch==2.9.1
torchvision>=0.18.0  # Compatible with torch 2.9.1

# ============================================================================
# NLP - Transformers
# ============================================================================
transformers==4.57.1
tokenizers==0.22.1

# ============================================================================
# NLP - Text Processing
# ============================================================================
spacy==3.8.9
nltk==3.9.1

# ============================================================================
# Time Series Processing
# ============================================================================
tsfresh==0.21.1

# ============================================================================
# Visualization
# ============================================================================
matplotlib==3.10.0
seaborn==0.13.2
plotly==5.24.1

# ============================================================================
# Explainability
# ============================================================================
shap==0.50.0

# ============================================================================
# Training & Monitoring
# ============================================================================
tensorboard==2.20.0
tqdm==4.67.1

# ============================================================================
# Utilities
# ============================================================================
pyyaml==6.0.2
jsonlines>=4.0.0

# ============================================================================
# Optional Dependencies (not required for core functionality)
# ============================================================================
# accelerate  # For distributed training
# pytorch-lightning  # Alternative training framework
# torchmetrics>=1.0.0  # Additional metrics
# sentence-transformers  # Alternative text embeddings
# captum  # Additional explainability
# lime  # Additional explainability
# alibi  # Additional explainability
# streamlit  # For web UI
# wandb  # Alternative experiment tracking
# lifelines  # Survival analysis
# textstat  # Text statistics
# tsai>=0.4.0  # Time series augmentation
# sktime>=0.21.0  # Additional time series tools

# ============================================================================
# Notes:
# ============================================================================
# - scispacy is not included due to compatibility issues with Python 3.13
#   The pipeline uses en_core_web_sm instead which works perfectly
# 
# - After installation, download required spaCy model:
#   python -m spacy download en_core_web_sm
#
# - For GPU support on NVIDIA cards, install CUDA-enabled PyTorch:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
#
# - For Apple Silicon Macs, use the standard installation (MPS is automatically supported)
